requirements.txt
streamlit
numpy-financial
psycopg2-binary
openpyxl
reportlab